refactor: legacy: group network_v1 files by protocol role
# Summary

The network_v1 module is reorganized into errors/, message/, and
receiver/ subdirectories with one type per file. Every public path is
unchanged.

# Details

Nine files sat in one flat directory with no hint in their names of which
side of the protocol they served: telling sender.rs, snapshot_receiver.rs
and streaming.rs apart meant opening them. The top level now shows only
what a user of the crate touches (the RaftNetwork trait, Adapter, and the
chunk Sender), with the receive-side state machinery grouped under
receiver/ and the wire messages and errors in modules of their own,
mirroring the errors/ and raft/message/ layout of the openraft crate.

Two files carried two public types each and are split accordingly:
InstallSnapshotRequest and InstallSnapshotResponse into message/, and
StreamingState out of streaming.rs. Where a file name did not match the
type it holds, the file is renamed after the type, so a type stays
locatable from its path.

Cross-module imports are unified on absolute crate::network_v1::* paths.
The previous mix of super:: and crate:: forms turns ambiguous once files
sit at different depths.

The re-export surface had to stay fixed: openraft's deprecation notes
direct users to openraft_legacy::network_v1::*, and the network-v1-http
example consumes the prelude. Confirmed by diffing rustdoc JSON for the
publicly reachable API against the parent commit; all 327 item and member
signatures are identical.
