+++ legacy/src/network_v1/errors/install_snapshot_error.rs
