+++ legacy/src/network_v1/errors/snapshot_mismatch.rs
