+++ legacy/src/network_v1/errors/snapshot_segment_id.rs
