+++ legacy/src/network_v1/message/mod.rs
@@ -1,0 +1,7 @@
+//! Request and response messages of the v1 chunk-based snapshot RPC.
+
+mod install_snapshot_request;
+mod install_snapshot_response;
+
+pub use install_snapshot_request::InstallSnapshotRequest;
+pub use install_snapshot_response::InstallSnapshotResponse;
